revert previous change

+++ server/db/models/Team.js
@@ -17,168 +17,168 @@ module.exports = db.define('team', {
     }
   },
   mp: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   fg: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   fga: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   fgp: {
-    type: Sequelize.DECIMAL(10, 3),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   thr: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   thrA: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   thrP: {
-    type: Sequelize.DECIMAL(10, 3),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   two: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   twoA: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   twoP: {
-    type: Sequelize.DECIMAL(10, 3),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ft: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ftA: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ftP: {
-    type: Sequelize.DECIMAL(10, 3),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   orb: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   drb: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   trb: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ast: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   stl: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   blk: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   tov: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   pf: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   pts: {
-    type: Sequelize.DECIMAL(10, 1),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   ortg: {
-    type: Sequelize.DECIMAL(10, 2),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
     }
   },
   drtg: {
-    type: Sequelize.DECIMAL(10, 2),
+    type: Sequelize.FLOAT,
     allowNull: false,
     validate: {
       notEmpty: true
